+++ Runtime/Attributes/UIPanelAttribute.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Feif.UIFramework
-{
-    [AttributeUsage(AttributeTargets.Class)]
-    public class UIPanelAttribute : Attribute
-    {
-    }
-}