+++ Runtime/Layers/PanelLayer.cs
@@ -1,0 +1,8 @@
+namespace Feif.UIFramework
+{
+    public class PanelLayer : UILayer
+    {
+        public override string GetName() => "PanelLayer";
+        public override int GetOrder() => 100;
+    }
+}