+++ Runtime/Layers/WindowLayer.cs
@@ -1,0 +1,8 @@
+namespace Feif.UIFramework
+{
+    public class WindowLayer : UILayer
+    {
+        public override string GetName() => "WindowLayer";
+        public override int GetOrder() => 200;
+    }
+}