+++ Samples~/Scripts/Demo2/Window/UIStarred_Demo2.cs
@@ -17,7 +17,7 @@ namespace Feif.UI
     }
 
     // 这是一个Window，不需要UIData则继承UIBase，需要UIData则继承UIComponent
-    [UIWindow]
+    [WindowLayer]
     public class UIStarred_Demo2 : UIBase
     {
         [SerializeField] private RectTransform content;